ERS-55: Refactored generateNotes
1. Use promise based API
2. Remove ETL rest calls in favour of direct API calls.
