+++ package.json
@@ -47,9 +47,11 @@
   },
   "devDependencies": {
     "chai": "^3.4.1",
+    "chai-as-promised": "^5.3.0",
     "mocha": "^2.3.3",
     "nock": "^7.2.2",
     "sinon": "^1.17.3",
+    "sinon-as-promised": "^4.0.0",
     "sinon-chai": "^2.8.0"
   },
   "repository": {
